requirements.txt
nose
Django==1.6.2
beautifulsoup4
python-dateutil
